% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-database.R
\name{connect_database}
\alias{connect_database}
\title{Connect to the database}
\usage{
connect_database(
  user = "admin",
  password = "p@ssw0rd",
  host = "localhost",
  port = 5432,
  dbname = "roadnetdb"
)
}
\arguments{
\item{user}{The user name (default: "admin").}

\item{password}{The password (default: "admin").}

\item{host}{The host name (default: "localhost").}

\item{port}{The port number (default: 5432).}

\item{dbname}{The database name (default: "roadnetdb").}
}
\value{
A connection to the database.
}
\description{
Connect to the database.
To disconnect, call \code{disconnect_database()}.
}
\examples{
\dontrun{
connect_database()
}
}
