% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-table-osm-roads.R
\name{create_table_osm_roads}
\alias{create_table_osm_roads}
\title{Create a table of OpenStreetMap road}
\usage{
create_table_osm_roads(conn)
}
\arguments{
\item{conn}{A \code{DBIConnection}, a connection object to the database.}
}
\value{
The function invisibly returns
\code{TRUE} if the table was successfully created, \code{FALSE} otherwise.
}
\description{
Create a table named \code{osm_roads} in the database.
The \code{osm_roads} manages OpenStreetMap road data and
has the following columns:
\itemize{
\item \code{osm_id}: The ID of the road in OpenStreetMap
\item \code{road_type}: The type of the road
\item \code{road_name}: The name of the road
\item \code{layer}: The layer of the road.
The layer represents the height of the road.
\item \code{oneway}: Whether the road is one-way or not
\item \code{geom}: The geometry of the road
}
}
\examples{
\dontrun{
create_table_osm_roads(con)
}
}
