% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect-database.R
\name{disconnect_database}
\alias{disconnect_database}
\title{Disconnect from the database}
\usage{
disconnect_database(conn)
}
\arguments{
\item{conn}{A \code{DBIConnection}, a connection object to the database.}
}
\value{
The function returns
\code{TRUE} if this is connected and disconnecting was done as expected or
\code{FALSE} if this is not connected to the database.
}
\description{
Disconnect from the database.
}
\examples{
\dontrun{
disconnect_database(con)
}
}
