% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-osm-roads.R
\name{download_osm_roads}
\alias{download_osm_roads}
\title{Download road data from OpenStreetMap}
\usage{
download_osm_roads(area)
}
\arguments{
\item{area}{A \code{sf}, area to be downloaded}
}
\value{
OpenStreetMap road data
}
\description{
The function download road data from OpenStreetMap using overpass query.
Roads are records have the following values for the key, 'highway'
in OpenStreetMap data.
\itemize{
\item motorway
\item trunk
\item primary
\item secondary
\item tertiary
\item unclassified
\item residential
\item motorway_link
\item trunk_link
\item primary_link
\item secondary_link
\item tertiary_link
}

For one-way roads, the column \code{oneway} is \code{TRUE}.
}
\examples{
\dontrun{
download_osm_roads(toyota)
}
}
