% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable-extension.R
\name{enable_extension}
\alias{enable_postgis}
\alias{enable_pgrouting}
\title{Enable database extensions.}
\usage{
enable_postgis(conn)

enable_pgrouting(conn)
}
\arguments{
\item{conn}{A \code{DBIConnection}, a connection object to the database.}
}
\value{
The function returns
\code{FALSE} if enabling fails or \code{TRUE} if enabling succeeds.
}
\description{
Enable database extensions.
}
\examples{
\dontrun{
enable_postgis(con)

enable_pgrouting(con)
}
}
