% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-road-network.R
\name{generate_road_network}
\alias{generate_road_network}
\title{Generate road network from road data}
\usage{
generate_road_network(
  conn,
  area,
  node_table = "road_nodes",
  edge_table = "road_edges"
)
}
\arguments{
\item{conn}{A \code{DBIConnection}, a connection object to the database.}

\item{area}{A \code{sf}, area to be computed.}

\item{node_table}{A \code{character}, the name of the road network node table
(default: \code{road_nodes}).}

\item{edge_table}{A \code{character}, the name of the road network edge table
(default: \code{road_edges}).}
}
\value{
Number of rows of data updated.
}
\description{
The function generates a road network in the specified area.
A road network has intersection as nodes and road segments as edges.
The function creates two tables representing edges and nodes of the network.

The function generates a road network by the following steps:
\enumerate{
\item Generate road network edges from road data
\item Generate road network nodes from road network edges
\item Store the road network edges and nodes into the database
}

The road network edge table has the following columns:
\itemize{
\item \code{id}: The ID of the edge
\item \code{source}: The ID of the source node
\item \code{target}: The ID of the target node
\item \code{oneway}: Whether the edge is one-way or not
\item \code{length}: The length of the edge in meters
\item \code{parent}: The ID of the original road in OpenStreetMap
\item \code{geom}: The geometry of the edge
}

The road network nodes table has the following columns:
\itemize{
\item \code{id}: The ID of the node
\item \code{geom}: The geometry of the node
}

The nodes are computed by \code{pgr_createTopology} function in pgRouting.
If the distance between two nodes is less than 1 meter
in Web Mercator (EPSG 3857), the two nodes are merged into one node.
}
\examples{
\dontrun{
generate_road_network(con, toyota)
}
}
