% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-segmentized-network.R
\name{generate_segmentized_network}
\alias{generate_segmentized_network}
\title{Generate segmentized network from road network edges}
\usage{
generate_segmentized_network(
  conn,
  source_table = "road_edges",
  node_table = "segment_nodes",
  edge_table = "segment_edges",
  segment_length = 10
)
}
\arguments{
\item{conn}{A \code{DBIConnection}, a connection object to the database.}

\item{source_table}{A \code{character}, the name of the road network edge table
(default: \code{road_edges}).}

\item{node_table}{A \code{character}, the name of the segmentized network
node table (default: \code{segment_nodes}).}

\item{edge_table}{A \code{character}, the name of the segmentized network
edge table (default: \code{segment_edges}).}

\item{segment_length}{A \code{numeric}, the length of segments in meters
(default: 10).}
}
\value{
Number of rows of data updated.
}
\description{
The function generates a segmentized network in the specified area.
A segmentized network is a network in which the edges of a road network
are divided into segments of a specified length.
The function creates two tables representing edges and nodes of the network.

The function generates a segmentized network by the following steps:
\enumerate{
\item Split road network edges into segments as segmentized network edges
\item Generate segmentized network nodes from segmentized network edges
\item Store the segmentized network edges and nodes into the database
}

The node and edge tables of the segmentized network have the same columns
as the road network node and edge tables.
}
\examples{
\dontrun{
generate_segmentized_network(con)
}
}
\seealso{
\code{\link{generate_road_network}}
}
