% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-database.R
\name{launch_database}
\alias{launch_database}
\title{Launch the database}
\usage{
launch_database(
  user = "admin",
  password = "p@ssw0rd",
  dbname = "roadnetdb",
  port = 5432
)
}
\arguments{
\item{user}{The user name (default: "admin").}

\item{password}{The password (default: "p@ssword").}

\item{dbname}{The database name (default: "roadnetdb").}

\item{port}{The port number (default: 5432).}
}
\value{
The return code of this process.
}
\description{
Launch the database using docker.
This function requires docker to be installed beforehand.
}
\examples{
\dontrun{
launch_database()
}
}
