% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-osm-roads.R
\name{register_osm_roads}
\alias{register_osm_roads}
\title{Register OpenStreetMap road data into the database}
\usage{
register_osm_roads(conn, osm_roads)
}
\arguments{
\item{conn}{A \code{DBIConnection}, a connection object to the database.}

\item{osm_roads}{A data frame of OpenStreetMap road data.}
}
\value{
Number of rows of data updated.
}
\description{
Register OpenStreetMap road data into the database.
If the specified road data has already been registered,
this function update the data by specified road data as new data.
}
\examples{
\dontrun{
toyota_roads <- read_osm_roads(toyota)
register_osm_roads(con, toyota_roads)
}
}
