% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadnetdb-package.R
\docType{package}
\name{roadnetdb-package}
\alias{roadnetdb}
\alias{roadnetdb-package}
\title{roadnetdb: Package for road network operation with pgRouting}
\description{
This package provides functions for road network operation with pgRouting.
}
\author{
\strong{Maintainer}: Keisuke ANDO \email{ando@maslab.aitech.ac.jp}

}
\keyword{internal}
