% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutdown-database.R
\name{shutdown_database}
\alias{shutdown_database}
\title{Shutdown the database}
\usage{
shutdown_database()
}
\value{
The return code of this process.
}
\description{
Shutdown the database using docker.
This function requires docker to be installed beforehand.
}
\examples{
\dontrun{
shutdown_database()
}
}
